tests/test_p1_api.py
#!/usr/bin/env python3
"""
Test P1 API implementation - validates basic REST API functionality.
"""

import asyncio
import json
import logging
import threading
import time

import requests

from api.server import AgentNetAPI

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


def start_api_server():
    """Start the API server in a separate thread."""
    api = AgentNetAPI()
    server_thread = threading.Thread(target=api.run, args=("127.0.0.1", 8080))
    server_thread.daemon = True
    server_thread.start()
    time.sleep(2)  # Give server time to start
    return api


def test_api_endpoints():
    """Test the API endpoints."""
    print("🧪 Testing P1 API Implementation...")

    # Start API server
    api = start_api_server()
    base_url = "http://127.0.0.1:8080"

    try:
        # Test 1: Health check
        print("  Testing health endpoint...")
        response = requests.get(f"{base_url}/health")
        assert response.status_code == 200
        health_data = response.json()
        assert health_data["status"] == "healthy"
        print("  ✅ Health endpoint working")

        # Test 2: Create session
        print("  Testing session creation...")
        session_request = {
            "topic": "AI ethics framework",
            "agents": [
                {
                    "name": "Ethicist",
                    "style": {"logic": 0.8, "creativity": 0.6, "analytical": 0.9},
                },
                {
                    "name": "Technologist",
                    "style": {"logic": 0.9, "creativity": 0.7, "analytical": 0.8},
                },
            ],
            "mode": "debate",
            "max_rounds": 3,
            "convergence": True,
            "parallel_round": False,
            "convergence_config": {
                "convergence_min_overlap": 0.4,
                "convergence_window": 2,
            },
        }

        response = requests.post(f"{base_url}/sessions", json=session_request)
        assert response.status_code == 200
        session_data = response.json()
        session_id = session_data["session_id"]
        assert session_data["status"] == "ready"
        assert len(session_data["participants"]) == 2
        print(f"  ✅ Session created: {session_id}")

        # Test 3: Get session status
        print("  Testing session status...")
        response = requests.get(f"{base_url}/sessions/{session_id}/status")
        assert response.status_code == 200
        status_data = response.json()
        assert status_data["session_id"] == session_id
        assert status_data["status"] == "ready"
        print("  ✅ Session status working")

        # Test 4: Run session
        print("  Testing session execution...")
        response = requests.post(f"{base_url}/sessions/{session_id}/run")
        assert response.status_code == 200
        result_data = response.json()
        assert result_data["status"] == "completed"
        assert "converged" in result_data
        assert "rounds_executed" in result_data
        print(
            f"  ✅ Session completed: converged={result_data['converged']}, rounds={result_data['rounds_executed']}"
        )

        # Test 5: Get completed session
        print("  Testing session retrieval...")
        response = requests.get(f"{base_url}/sessions/{session_id}")
        assert response.status_code == 200
        full_session = response.json()
        assert full_session["status"] == "completed"
        assert full_session["converged"] is not None
        print("  ✅ Session retrieval working")

        # Test 6: List sessions
        print("  Testing session listing...")
        response = requests.get(f"{base_url}/sessions")
        assert response.status_code == 200
        sessions_list = response.json()
        assert "sessions" in sessions_list
        assert len(sessions_list["sessions"]) >= 1
        print(
            f"  ✅ Session listing working: {len(sessions_list['sessions'])} sessions"
        )

        print("\n🎉 All P1 API tests passed!")
        return True

    except Exception as e:
        print(f"\n❌ API test failed: {e}")
        return False


def test_api_without_server():
    """Test API functionality without running HTTP server."""
    print("🧪 Testing P1 API Core Functionality...")

    api = AgentNetAPI()

    # Test session creation
    request_data = {
        "topic": "Sustainable technology",
        "agents": [
            {
                "name": "Engineer",
                "style": {"logic": 0.9, "creativity": 0.5, "analytical": 0.8},
            },
            {
                "name": "Environmentalist",
                "style": {"logic": 0.7, "creativity": 0.8, "analytical": 0.6},
            },
        ],
        "mode": "consensus",
        "max_rounds": 4,
        "convergence": True,
        "parallel_round": True,
        "convergence_config": {"convergence_min_overlap": 0.3, "convergence_window": 3},
    }

    session_response = api.create_session(request_data)
    session_id = session_response["session_id"]

    print(f"  ✅ Session created: {session_id}")
    print(f"  ✅ Participants: {session_response['participants']}")

    # Test session status
    status = api.get_session_status(session_id)
    print(f"  ✅ Initial status: {status['status']}")

    # Test session execution
    async def run_test():
        result = await api.run_session(session_id)
        print(f"  ✅ Session completed: {result['status']}")
        print(
            f"  ✅ Converged: {result['converged']}, Rounds: {result['rounds_executed']}"
        )
        return result

    result = asyncio.run(run_test())

    # Test final session state
    final_session = api.get_session(session_id)
    print(f"  ✅ Final session status: {final_session['status']}")

    print("\n🎉 All P1 API core tests passed!")
    return True


if __name__ == "__main__":
    print("P1 API Implementation Tests")
    print("=" * 50)

    # Test core functionality first (doesn't require HTTP server)
    core_success = test_api_without_server()

    print("\n" + "=" * 50)

    # Test HTTP endpoints (requires server)
    try:
        http_success = test_api_endpoints()
    except Exception as e:
        print(f"HTTP API tests skipped due to: {e}")
        http_success = False

    print("\n" + "=" * 50)
    if core_success:
        print("🏆 P1 API Core Implementation: COMPLETE")
    if http_success:
        print("🏆 P1 API HTTP Endpoints: COMPLETE")

    print("\nP1 Basic API Foundation Successfully Implemented!")
